requirements.txt
# Core simulation
robosuite>=1.4.0
mujoco>=3.0.0
gymnasium>=0.29.0
numpy>=1.24.0

# Training
torch>=2.0.0
stable-baselines3>=2.0.0
wandb

# Dev
pytest>=7.0
pytest-cov
pyyaml
